% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_fasta_functions_v5.R
\name{get_truncated_fasta_file}
\alias{get_truncated_fasta_file}
\title{get_truncated_fasta_file}
\usage{
get_truncated_fasta_file(ECF_name, genome_fasta_filenames, annotation_file,
  annotation_file2, annotation_file3, how_much_seq = c(-332, 1))
}
\arguments{
\item{ECF_name}{a ProGenomes gene name.}

\item{genome_fasta_filenames}{a string describing the location of a fasta file.}

\item{annotation_file}{either a string or a data frame containing the primary annotation for the genome.}

\item{annotation_file2}{either a string or a matrix containing the secondary annotation for the genome.}

\item{annotation_file3}{either a string or a matrix containing the tertiary annotation for the genome.}

\item{how_much_seq}{a vector of containing the start and stop off region of interest upstream of a gene. Defaults to c(-332,1).}
}
\description{
This function takes in a file name and some annotations. It then chops up the genome into just the bits upstream of genes.
The function first reads in the genome, then chops it up (padding with n's)
Returns a list containing [[1]] chopped up genome [[2]] corresponding annotation [[3]] the value of where the genome was chopped.
}
\examples{
get_truncated_fasta_file("1114970.PRJNA76715.PSF113_1768", 
"~/Documents/eggNOG ECF pipeline/eggNOG split files/Proteobacteria/contigs/1114970.PRJNA76715.contigs.fasta",
"~/Documents/eggNOG ECF pipeline/eggNOG split files/Proteobacteria/annot/1114970.PRJNA76715.annot.tsv" ,
"~/Documents/eggNOG ECF pipeline/eggNOG split files/Proteobacteria/annot2/1114970.PRJNA76715.annot2.tsv" ,
"~/Documents/eggNOG ECF pipeline/eggNOG split files/Proteobacteria/annot3/1114970.PRJNA76715.annot3.tsv" , c(-332, 1))
}
\keyword{genome}
